# optimize crc32c_hw_series()