(1) reimplement crc32c_hw_series() making it more friendly to cache
(2) add perf routines and compare it with naive implementation

results on MBP with M1 @ 3.2GHz:
```
[ RUN      ] Perf.crc32c_hw_series
crc32c_hw_series (131072 bytes * 81920 rounds = 10 GB), time spent: 432455 us (23.12 GB/s)
crc32c_hw_series (1048576 bytes * 10240 rounds = 10 GB), time spent: 429289 us (23.29 GB/s)
crc32c_hw_series_naive (131072 bytes * 81920 rounds = 10 GB), time spent: 456051 us (21.93 GB/s)
crc32c_hw_series_naive (1048576 bytes * 10240 rounds = 10 GB), time spent: 484315 us (20.65 GB/s)
[       OK ] Perf.crc32c_hw_series (1802 ms)
```

results on a x86 server with Intel(R) Xeon(R) Silver 4114 CPU @ 2.20GHz:
```
[ RUN      ] Perf.crc32c_hw_series
crc32c_hw_series (131072 bytes * 81920 rounds = 10 GB), time spent: 619618 us (16.14 GB/s)
crc32c_hw_series (1048576 bytes * 10240 rounds = 10 GB), time spent: 617740 us (16.19 GB/s)
crc32c_hw_series_naive (131072 bytes * 81920 rounds = 10 GB), time spent: 764477 us (13.08 GB/s)
crc32c_hw_series_naive (1048576 bytes * 10240 rounds = 10 GB), time spent: 768718 us (13.01 GB/s)
[       OK ] Perf.crc32c_hw_series (2771 ms)
```